+++ sketch-core/src/androidMain/kotlin/com/github/panpf/sketch/state/DrawableStateImage.kt
@@ -16,6 +16,7 @@
 
 package com.github.panpf.sketch.state
 
+import android.graphics.drawable.ColorDrawable
 import android.graphics.drawable.Drawable
 import androidx.annotation.DrawableRes
 import com.github.panpf.sketch.Image
@@ -25,6 +26,7 @@ import com.github.panpf.sketch.drawable.DrawableFetcher
 import com.github.panpf.sketch.drawable.EquitableDrawable
 import com.github.panpf.sketch.drawable.RealEquitableDrawable
 import com.github.panpf.sketch.drawable.ResDrawable
+import com.github.panpf.sketch.drawable.asEquitable
 import com.github.panpf.sketch.request.ImageRequest
 
 /**
@@ -33,6 +35,22 @@ import com.github.panpf.sketch.request.ImageRequest
  * @see com.github.panpf.sketch.core.android.test.state.DrawableStateImageTest.testAsStateImage
  */
 fun EquitableDrawable.asStateImage(): DrawableStateImage = DrawableStateImage(this)
+
+/**
+ * [Drawable] as [StateImage]
+ *
+ * @see com.github.panpf.sketch.core.android.test.state.DrawableStateImageTest.testAsStateImage
+ */
+fun Drawable.asStateImage(equalityKey: Any): DrawableStateImage =
+    DrawableStateImage(RealEquitableDrawable(this.asEquitable(equalityKey)))
+
+/**
+ * [ColorDrawable] as [StateImage]
+ *
+ * @see com.github.panpf.sketch.core.android.test.state.DrawableStateImageTest.testAsStateImage
+ */
+fun ColorDrawable.asStateImage(): DrawableStateImage =
+    DrawableStateImage(RealEquitableDrawable(this.asEquitable()))
 
 /**
  * Create a [DrawableStateImage] with [EquitableDrawable]